minor error handling etc.

+++ SimpleShell.c
@@ -11,7 +11,7 @@
 #include <semaphore.h>
 #include <fcntl.h>
 
-#define SHM_NAME "/processtable"
+#define SHM_NAME "/processtable_shm"
 
 typedef struct{
     char command[1024];
@@ -25,7 +25,7 @@ typedef struct{
 typedef struct{
     Process table[256];
     int count;
-    sem_t sem; 
+    sem_t sem;
 } SharedProcessTable;
 
 SharedProcessTable* processtable;
@@ -38,9 +38,9 @@ int length(int n) {
         return 1;
     }
     int l = 0;
-    n = abs(n); 
+    n = abs(n);
     while (n > 0) {
-        n /= 10; // 
+        n /= 10;
         l++;
     }
     return l;
@@ -83,19 +83,28 @@ static void my_handler(int signal){
 
 void submit(char* command){
 	char** args = read_command(command);//read arguments
+	if (args == NULL){
+		return;
+	}
 	if (strcmp(args[0],"submit") == 0){
-		printf("%s",args[1]);
+		if (!args[1]){
+			printf("Usage: submit ./a.out");//insufficent args
+			free(args);
+			return;
+		}
 		//executable submitted
 		//fork to create child process & execute - execution pauses due to dummy main calling raise(SIGSTOP)
 		pid_t child = fork();
 		if (child == 0){
 			//in child process
-			execvp(args[1],++args);//++args to skip the submit word
+			execvp(args[1],&args[1]);//++args to skip the submit word
 			perror("exec failed");
 			exit(1);
 		}
 		else if (child > 0){
 			//in parent
+			int status;
+			waitpid(child,&status,WUNTRACED);//wait for child to raise SIGSTOP
 			sem_wait(&processtable->sem);//modifying in shm
 			if (processtable->count < 256){
 				Process* p = &(processtable->table[processtable->count]);
@@ -105,6 +114,7 @@ void submit(char* command){
 				p->state = 0; //running
 				strcpy(p->command,args[1]);
 				processtable->count++;
+				printf("Process added!! PID: %d, Name: %s\n",child,args[1]);
 			//parent adds process to process table in shared memory using semaphores
 			}
 			else{
@@ -113,17 +123,18 @@ void submit(char* command){
 			sem_post(&processtable->sem);
 		}
 	}
-	else{
-		//shell need not handle other commands for this assignment
-		return;
-	}
+
+	//shell need not handle other commands for this assignment
+
+	free(args);
+	return;
 }
 
 
 void display_info(){
 	int c=length(processtable->count);
 	if (c>3){
-		printf("%*s", c-3);
+		printf("%*s", c-3," ");
 	}
 	printf("PID   Wait time(ms)   Completion Time(ms)   Command\n");
 	for (int i = 0; i < processtable->count; i++){
@@ -155,43 +166,54 @@ void display_info(){
 }
 
 void setup_shm_sem(){
-    //shm setup
-    shm_fd = shm_open(SHM_NAME, O_CREAT | O_RDWR, 0666);
-    if (shm_fd < 0){
-        perror("shm_open failed");
-        exit(1);
-    }
-    ftruncate(shm_fd, sizeof(SharedProcessTable));
-    processtable = mmap(NULL, sizeof(SharedProcessTable), PROT_READ | PROT_WRITE, MAP_SHARED, shm_fd, 0);
-    if (processtable == MAP_FAILED){
-        perror("mmap failed");
-        exit(1);
-    }
-    processtable->count = 0;
+    	//shm setup
+    	shm_fd = shm_open(SHM_NAME, O_CREAT | O_RDWR, 0666);
+    	if (shm_fd < 0){
+        	perror("shm_open failed");
+        	exit(1);
+    	}
+    	if (ftruncate(shm_fd, sizeof(SharedProcessTable)) == -1){
+		perror("ftruncate failed");
+		exit(1);
+	}
+    	processtable = mmap(NULL, sizeof(SharedProcessTable), PROT_READ | PROT_WRITE, MAP_SHARED, shm_fd, 0);
+    	if (processtable == MAP_FAILED){
+        	perror("mmap failed");
+        	exit(1);
+    	}
+    	processtable->count = 0;
 
-    //semaphore setup
-    sem_init(&processtable->sem, 1, 1);
+    	//semaphore setup
+    	if (sem_init(&processtable->sem, 1, 1) == -1){
+		perror("sem_init failed");
+		exit(1);
+	}
 }
 
 
 void cleanup(){
 	//cleanup semaphore, shared memory
-    sem_destroy(&processtable->sem);
-    munmap(processtable, sizeof(SharedProcessTable));
-    close(shm_fd);
-    shm_unlink("/processtable");
+    	sem_destroy(&processtable->sem);
+    	munmap(processtable, sizeof(SharedProcessTable));
+    	close(shm_fd);
+    	shm_unlink("SHM_NAME");
 }
 
 void shell_loop(){
 	while (1) {
-		printf("group60@os:~$");
+		printf("group60@os:~$ ");
 		char* command = (char*)malloc(1024*sizeof(char)); //1KB input
+		if (!command){
+			perror("malloc failed");
+			exit(1);
+		}
 		fgets(command,1024,stdin);
 		if (strcmp(command,"exit\n") == 0){
 			//terminate scheduler by sending sigint (scheduler has a sigint handler to wait for each process to end, display details & then exit)
 			kill(scheduler_pid, SIGINT);
 			waitpid(scheduler_pid,NULL,0);
 			display_info();
+			free(command);
 			//exit the shell
 			break;
 		}
@@ -203,49 +225,50 @@ void shell_loop(){
 }
 
 int main(int argc,char** argv){
-    if (argc != 3){
-        fprintf(stderr, "Usage: ./SimpleShell ncpu tslice\n");
-        exit(1);
-    }
-    int ncpu;
-    ncpu = atoi(argv[1]);
-    tslice = atoi(argv[2]);
-    if (ncpu < 1 || tslice < 1){
-        fprintf(stderr, "Non-positive ncpu or tslice\n");
-        exit(1);
-    }
-    
-    setup_shm_sem();
-    
-    //signal handler for SIGCHLD
-    struct sigaction sa;
-    memset(&sa, 0, sizeof(sa));
-    sa.sa_handler = my_handler;
-    sigaction(SIGCHLD, &sa, NULL);
+    	if (argc != 3){
+        	perror("Usage: ./SimpleShell ncpu tslice\n");
+        	exit(1);
+    	}
+    	int ncpu;
+    	ncpu = atoi(argv[1]);
+    	tslice = atoi(argv[2]);
+    	if (ncpu < 1 || tslice < 1){
+        	perror("Non-positive ncpu or tslice\n");
+        	exit(1);
+	}
 
-    //process for scheduler
-    scheduler_pid = fork();
-    if (scheduler_pid == 0){
-        //execute the scheduler program using execvp
-        char ncpu_str[10], tslice_str[10];
-        sprintf(ncpu_str, "%d", ncpu);
-        sprintf(tslice_str, "%d", tslice);
-        
-        char* scheduler_args[] = {"./SimpleScheduler", ncpu_str, tslice_str, NULL};
-        execvp(scheduler_args[0], scheduler_args);
-        
-        //only runs if execvp fails
-        perror("execvp for scheduler failed");
-        exit(1);
-    }
-    else if (scheduler_pid < 0){
-        perror("fork for scheduler failed");
-        exit(1);
-    }
+    	setup_shm_sem();
 
-    shell_loop();
-    
-    cleanup();
-    
-    return 0;
+    	//signal handler for SIGCHLD
+    	struct sigaction sa;
+    	memset(&sa, 0, sizeof(sa));
+	sa.sa_flags = SA_RESTART;
+    	sa.sa_handler = my_handler;
+	sigemptyset(&sa.sa_mask);
+    	sigaction(SIGCHLD, &sa, NULL);
+
+    	//process for scheduler
+    	scheduler_pid = fork();
+    	if (scheduler_pid == 0){
+        	//execute the scheduler program using execvp
+        	char ncpu_str[10], tslice_str[10];
+        	sprintf(ncpu_str, "%d", ncpu);
+	        sprintf(tslice_str, "%d", tslice);
+
+        	char* scheduler_args[] = {"./SimpleScheduler", ncpu_str, tslice_str, NULL};
+		setpgid(0, 0);
+	        execvp(scheduler_args[0], scheduler_args);
+
+        	//only runs if execvp fails
+        	perror("execvp for scheduler failed");
+        	exit(1);
+    	}
+    	else if (scheduler_pid < 0){
+        	perror("fork for scheduler failed");
+        	exit(1);
+    	}
+
+	shell_loop();
+	cleanup();
+	return 0;
 }
